requirements.txt
pjsua2==2.13
aiohttp==3.9.1
websockets==12.0
pydub==0.25.1
pyaudio==0.2.11
requests==2.31.0
asyncio==3.4.3
numpy==1.24.3
soundfile==0.12.1
python-dotenv==1.0.0
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0 